+++ packages/serialize-query-params/src/__tests__/helpers.ts
@@ -1,7 +1,7 @@
 import { stringify } from 'query-string';
 import { EncodedQuery } from '..';
 
-export function makeMockLocation(query: EncodedQuery): Location {
+export function makeMockLocation(query: EncodedQuery, includeHref: boolean = true): Location {
   const queryStr = stringify(query);
   const search = queryStr.length ? `?${queryStr}` : '';
   return {
@@ -9,7 +9,7 @@ export function makeMockLocation(query: EncodedQuery): Location {
     host: 'localhost:3000',
     pathname: '/',
     search,
-    href: 'http://localhost:3000/' + search,
+    href: includeHref ? 'http://localhost:3000/' + search : undefined,
     key: `mock-${Date.now()}`,
   } as Location & { key: string };
 }